Renamed matrx_core back to matrx
