--- __init__.py
@@ -1,4 +1,0 @@
-import matrx_core as matrx
-import matrx_visualizer
-
-
