+++ matrx/scenarios/new_vis_test.py
@@ -1,7 +1,7 @@
-from matrx_core.agents.patrolling_agent import PatrollingAgentBrain
-from matrx_core.world_builder import WorldBuilder
-from matrx_core.agents.human_agent_brain import HumanAgentBrain
-from matrx_core.actions.move_actions import *
+from matrx.agents.patrolling_agent import PatrollingAgentBrain
+from matrx.world_builder import WorldBuilder
+from matrx.agents.human_agent_brain import HumanAgentBrain
+from matrx.actions.move_actions import *
 
 
 def create_factory():
