+++ setup.py
@@ -8,7 +8,7 @@ setuptools.setup(
     name="matrx",
     version="0.0.93",
     author="The MATRX Development team at TNO.nl",
-    author_email="info@matrx_core-software.com",
+    author_email="info@matrx-software.com",
     description="A Python package for the rapid development of autonomous systems and human-agent teaming concepts.",
     long_description=long_description,
     long_description_content_type="text/markdown",
